Move network file into own module
