+++ doc/api/network.rst
@@ -1,0 +1,7 @@
+etrago.network module
+-------------------
+
+.. automodule:: etrago.network
+    :members:
+    :undoc-members:
+    :show-inheritance:
